% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ch_reco}
\alias{ch_reco}
\title{Generates a cohort reconstruction for salmon or other anadromous fish.}
\usage{
ch_reco(
  rel,
  reco,
  birth_month,
  size_at_age = length_at_age,
  rel_mort = release,
  nat_mort = nat_mort_default,
  sex = "both",
  fisheries = release,
  bootstrap = TRUE,
  iter = 1000,
  min_harvest_rate = 0,
  detail = T,
  level = 0.05,
  hpd = T
)
}
\value{
list of cohort reconstructions with or without bootstrapping.
}
\description{
Generates a cohort reconstruction for salmon or other anadromous fish.
}
