% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwtCensus-package.R
\docType{package}
\name{cwtCensus-package}
\alias{cwtCensus}
\alias{cwtCensus-package}
\title{cwtCensus: Efficient Population Analysis Tool for Cohort Reconstruction}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/J-Ch-n/cohort_reco_pk}
  \item Report bugs at \url{https://github.com/J-Ch-n/cohort_reco_pk/issues}
}

}
\keyword{internal}
