% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_at_age}
\alias{length_at_age}
\title{Length at age, or key length age, information}
\format{
\subsection{\code{length_at_age}}{

\describe{
\item{"age"}{Integer. The fishing age of the cohort.}
\item{"month"}{Integer. The month in numeric form. For instance, 3L is March.}
\item{"mean"}{Double. The mean total body length in inches.}
\item{"sd"}{Double. The standard deviation of total body length in inches.}
}
}
}
\usage{
length_at_age
}
\description{
Length at age, or key length age, information
}
\keyword{datasets}
