% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_mort_default}
\alias{nat_mort_default}
\title{Natural mortality information}
\format{
\subsection{'release'}{

\describe{
\item{age}{The age of fish in which the CWT is recovered}
\item{month}{The month in which the CWTs are released}
\item{rate}{The natural mortality rate for the specified age and month}
}
}
}
\usage{
nat_mort_default
}
\description{
Natural mortality information
}
\keyword{datasets}
