% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recovery}
\alias{recovery}
\title{Recovery information}
\format{
\subsection{'release'}{

\describe{
\item{run_year}{The year in which the CWT is recovered}
\item{recovery_id}{The recovery ID of the CWT}
\item{fishery}{The index of the fishery from which the CWT is recovered}
\item{tag_code}{The unique identification for each batch of CWT}
\item{length}{The length of tagged fish in inches}
\item{sex}{The biological sex of the tagged fish}
\item{month}{The month in which the CWT is recovered}
\item{location}{The code of the location from which the CWT is recovered}
\item{size_limit}{The minimum size limit for fish in inches at the time of recovery}
\item{est_num}{The estimated number of fish that each CWT represents.}
}
}
}
\usage{
recovery
}
\description{
Recovery information
}
\keyword{datasets}
