% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{release}
\alias{release}
\title{Release information}
\format{
\subsection{'release'}{

\describe{
\item{release_month}{The month in which the CWTs are released}
\item{brood_year}{The birth year of the tagged fish}
\item{tag_code}{The unique identification for each batch of CWT.}
\item{prod_exp}{The production expansion rate for each batch of CWT.}
\item{total_release}{The total number of fish released under a specific CWT ID}
}
}
}
\usage{
release
}
\description{
Release information
}
\keyword{datasets}
