% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{release_mort}
\alias{release_mort}
\title{Release mortality information}
\format{
\subsection{'release'}{

\describe{
\item{run_year}{The year in which the CWT is recovered}
\item{fishery}{The index of the fishery from which the CWT is recovered}
\item{location}{The code of the location from which the CWT is recovered}
\item{month}{The month in which the CWT is recovered}
\item{rate}{The release mortality rate for the specified location and time}
}
}
}
\usage{
release_mort
}
\description{
Release mortality information
}
\keyword{datasets}
