% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival_default}
\alias{survival_default}
\title{Age-specific survival information.}
\format{
\subsection{\code{survival_default}}{

\describe{
\item{age}{Integer. The age of fish in which the CWT is recovered}
\item{month}{Integer. The month in which the CWTs are released}
\item{rate}{Double. The natural mortality rate for the specified age and month}
}
}
}
\usage{
survival_default
}
\description{
Age-specific survival information.
}
\keyword{datasets}
