% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ch_reco}
\alias{ch_reco}
\title{Generates a cohort reconstruction for salmon or other anadromous fish.}
\usage{
ch_reco(
  rel,
  reco,
  size_at_age = release,
  rel_mort = release,
  nat_mort = release,
  sex = "both",
  fisheries = release,
  bootstrap = TRUE,
  iter = 1000
)
}
\value{
list of cohort reconstructions with or without bootstrapping.
}
\description{
Generates a cohort reconstruction for salmon or other anadromous fish.
}
