% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cohort_reconstruct}
\alias{cohort_reconstruct}
\title{Conduct cohort reconstruction for populations with coded wire tags}
\usage{
cohort_reconstruct(
  rel,
  reco,
  birth_month,
  last_month,
  fisheries = list(oc_rec = 40, oc_com = 10, esc_sp = 54, esc_hat = 50, riv_harv = 46),
  size_at_age = length_at_age,
  rel_mort = release_mort,
  survival = survival_default,
  d_mort = 0.05,
  hr_mort_com = 0.26,
  hr_mort_rec = 0.14,
  bootstrap = TRUE,
  iter = 1000,
  min_harvest_rate = 0,
  detail = TRUE,
  level = 0.05,
  hpd = TRUE,
  verbose = TRUE,
  sex = "both"
)
}
\arguments{
\item{rel}{Input data frame for CWT releases with columns:
\describe{
\item{brood_year}{Integer. The birth year of each CWT batch.}
\item{total_release}{Double. The total number of releases for each batch of CWT.}
\item{prod_exp}{Double. The ratio between the number of tagged fish and the total number of fish in each batch, i.e., number of tagged fish / batch size.}
\item{tag_code}{Any type. The identifier for each batch of CWT release.}
}}

\item{reco}{Input data frame for CWT recoveries with columns:
\describe{
\item{run_year}{Integer. The year of each CWT recovery.}
\item{month}{Integer. The month of each CWT recovery.}
\item{length}{Double. The length in inches of each recovered fish.}
\item{tag_code}{Any type. The identifier for each recovered tag.}
\item{fishery}{Double. The identifier for the fishery from which the tag is recovered. The fisheries include river harvest, spawning ground escapement, hatchery escapement, ocean recreational fishing, ocean commercial fishing.}
\item{location}{Character. The identifier for the location at which the tag is recovered.}
\item{size_limit}{Double. The minimum harvest limit (inclusive) in inches for the recovered tag's region and date.}
\item{est_num}{Double. The estimated number of fish each tag represents.}
\item{sex}{Character. The biological sex of the recovered fish.}
}}

\item{birth_month}{Integer specifying the birth month of all individuals in the data set.}

\item{fisheries}{Named list associating each type of fishery with a unique identifier in the release and recovery data. Must contain the following named elements:
\describe{
\item{oc_rec}{Identifier for recreational ocean fishery.}
\item{oc_com}{Identifier for commercial ocean fishery.}
\item{esc_sp}{Identifier for escapement to spawning ground.}
\item{esc_hat}{Identifier for escapement to hatchery.}
\item{riv_harv}{Identifier for river harvest.}
}}

\item{size_at_age}{Data frame specifying the mean and standard deviation for the individual body length at each age and month.}

\item{rel_mort}{Data frame specifying the release mortality rate due to ocean fishing in each region.
\describe{
\item{run_year}{Integer for the run year.}
\item{fishery}{Identifier for the fishery type.}
\item{location}{Identifier for location.}
\item{month}{Integer for the month.}
\item{rate}{Double for the release mortality rate.}
}}

\item{survival}{Data frame specifying the age specific natural survival rates.
\describe{
\item{age}{Numeric specifying the age.}
\item{rate}{Double specifying the agely survival rate.}
}}

\item{bootstrap}{Boolean indicating if parametric bootstrapping should be conducted. If \code{bootstrap} is set to false, then \code{iter} is ignored.}

\item{iter}{Numeric or integer specifying the number of iterations for parametric bootstrapping. The default value is 1000. Note that larger values require more computing power and memory.}

\item{min_harvest_rate}{Double specifying the lower bound for the harvest rate considered in cohort reconstruction.}

\item{detail}{Boolean indicating if the reconstruction should calculate breakdowns and summary information.}

\item{level}{Double between 0 and 1 inclusive, specifying the credible level for credible intervals. This option only matters if \code{bootstrap} is set to TRUE.}

\item{hpd}{Boolean indicating if the highest posterior density credible interval should be used. If FALSE, a symmetric credible interval is used.}

\item{sex}{String specifying which sex or sexes to consider. Must choose from "male", "female", or "both". Here, "male" is shorthand for sperm-producing individuals. "Female" is shorthand for egg-producing individuals.}
}
\value{
A three-dimensional list of results. The first dimension encodes brood year information, the second age, and the third month. Depending on \code{detail} and \code{bootstrap}, the corresponding output will vary.

\if{html}{\out{<div class="sourceCode">}}\preformatted{    by  age  month

├── [[2002]]
│   ├── [[2]]
│   │   ├── [[1]]
│   │   │   ├── `data`
│   │       └── `summary`
│   ├── [[3]]
│   │   ├── [[1]]
│   │   │   ├── `data`
│   │       └── `summary`
}\if{html}{\out{</div>}}
}
\description{
\code{cohort_reconstruct} uses released and recovered coded wire tags to calculate the ocean
abundance of the population for each brood year at each age and in each month.
The function returns a list of outputs indexed by brood year, age, and month. Each
element of the list contains two elements, \code{data} and \code{summary}. In general, the \code{data} segment
contains the raw data of cohort reconstruction, while the \code{summary} segment includes
statistics from the raw data.

Depending on if the \code{bootstrap} and \code{detail} flags are set, the \code{data} and \code{summary}
segments may behave in different ways. For more information, see details.
}
\details{
\subsection{Aging convention}{

We use fishing age, or year of life, as the age for captured tags. This aging convention starts from age-1. Namely,
as soon as the fish is born, it is in the first year of life, thus age-1. Accordingly, the age in the \code{size_at_age} data frame should use
fishing age as well.
}

\subsection{Detail, bootstrap, and results}{

Since \code{detail} and \code{bootstrap} are Boolean valued variables, there are four combinations these two variables.
\describe{
\item{\verb{detail = True, bootstrap = True}}{Provides results with detailed summary statistics using bootstrapped data.}
\item{\verb{detail = False, bootstrap = True}}{Provides results with cohort data only using bootstrapped data.}
\item{\verb{detail = True, bootstrap = False}}{Provides results with detailed summary statistics using point estimates.}
\item{\verb{detail = False, bootstrap = True}}{Provides results with cohort data only using point estimates.}
}
}

\subsection{Danger of incomplete cohort reconstruction}{

If the provided recovery data don't span the entire life time of a stock, cohort reconstruction may fail to
yield accurate information. Be sure to provide a complete cohort recovery data set.
}
}
\examples{
cohort_reconstruct(release, recovery, birth_month = 4,
  bootstrap = F, last_month = 12, iter = 10,
  level = 0.95, detail = F, sex = "both")
}
