% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cohort_reconstruct}
\alias{cohort_reconstruct}
\title{Conduct cohort reconstruction for populations with coded wire tags}
\usage{
cohort_reconstruct(
  rel,
  reco,
  birth_month,
  last_month = 12L,
  fisheries = list(oc_rec = 40, oc_com = 10, esc_sp = 54, esc_hat = 50, riv_harv = 46),
  size_at_age = length_at_age,
  rel_mort = release_mort,
  survival = survival_default,
  d_mort = 0.05,
  hr_mort_com = 0.26,
  hr_mort_rec = 0.14,
  bootstrap = TRUE,
  iter = 1000,
  detail = TRUE,
  min_harvest_rate = 0,
  level = 0.05,
  hpd = TRUE,
  verbose = TRUE,
  sex = "both"
)
}
\arguments{
\item{rel}{Input data frame for CWT releases with columns:
\describe{
\item{"brood_year"}{Integer. The birth year of each CWT batch.}
\item{"total_release"}{Double. The total number of releases for each batch of CWT.}
\item{"prod_exp"}{Double. The ratio between the number of tagged fish and the total number \cr
of fish in each batch, i.e., number of tagged fish / batch size.}
\item{"tag_code"}{Any type. The identifier for each batch of CWT release.}
}}

\item{reco}{Input data frame for CWT recoveries with columns:
\describe{
\item{"run_year"}{Integer. The year of each CWT recovery.}
\item{"month"}{Integer. The month of each CWT recovery.}
\item{"length"}{Double. The length in inches of each recovered fish.}
\item{"tag_code"}{Any type. The identifier for each recovered tag.}
\item{"fishery"}{Double. The identifier for the fishery from which the tag is recovered. \cr
The fisheries include river harvest, spawning ground escapement, hatchery \cr
escapement, ocean recreational fishing, ocean commercial fishing.}
\item{"location"}{Character. The identifier for the location at which the tag is recovered.}
\item{"size_limit"}{Double. The minimum harvest limit (inclusive) in inches for the recovered \cr
tag's region and date.}
\item{"est_num"}{Double. The estimated number of fish each tag represents.}
\item{"sex"}{Character. The biological sex of the recovered fish.}
}}

\item{birth_month}{Integer specifying the birth month of all individuals in the data set.}

\item{last_month}{Integer indicating the last month, in numeric form, that counts toward the \cr
current run. This variable is to separate late spawners of the current run and early spawners \cr
of the next run.
The default value is 12L, meaning the entire natural year counts toward the current run.}

\item{fisheries}{Named list associating each type of fishery with a unique identifier in the \cr
release and recovery data. Must contain the following named elements:
\describe{
\item{"oc_rec"}{Identifier for recreational ocean fishery. The default value is 40.}
\item{"oc_com"}{Identifier for commercial ocean fishery. The default value is 10.}
\item{"esc_sp"}{Identifier for escapement to spawning ground. The default value is 54.}
\item{"esc_hat"}{Identifier for escapement to hatchery. The default value is 50.}
\item{"riv_harv"}{Identifier for river harvest. The default value is 46.}
}}

\item{size_at_age}{Data frame specifying the mean and standard deviation for the individual \cr
body length at each age and month. The default data frame is included in the package and named \cr
\code{length_at_age}.
The following is a structure for the required data frame.
\describe{
\item{"age"}{Integer. The fishing age of the cohort.}
\item{"month"}{Integer. The month in numeric form. For instance, 3L is March.}
\item{"mean"}{Double. The mean total body length in inches.}
\item{"sd"}{Double. The standard deviation of total body length in inches.}
}}

\item{rel_mort}{Data frame specifying the release mortality rate due to ocean fishing in each \cr
region. The default data frame is included in the package and named \code{release_mort}. \cr
The following is a structure for the required data frame.
\describe{
\item{"run_year"}{Integer. The year of each CWT recovery.}
\item{"fishery"}{Any type. The identifier for each type of fishery.}
\item{"location"}{Any type. The identifier for location.}
\item{"month"}{Integer. The month of each CWT recovery.}
\item{"rate"}{Double. The release mortality rate for each timestep.}
}}

\item{survival}{Data frame specifying the age specific natural survival rates. The default data \cr
frame is included in the package and named \code{survival_default}. \cr
The following is a structure for the required data frame.
\describe{
\item{"age"}{Numeric specifying the age.}
\item{"rate"}{Double specifying the agely survival rate.}
}}

\item{d_mort}{Double indicating the drop off mortality of fishery impact. Drop off mortality is \cr
the chance that a fish dies after being hooked but not landed. The default value is 0.05. \if{html}{\out{<Source>}}}

\item{hr_mort_com}{Double indicating the catch and release mortality of commercial fishery. This \cr
is the chance that a fish dies after being landed and released in a commercial fishery. The \cr
default value is 0.26. \if{html}{\out{<Source>}}}

\item{hr_mort_rec}{Double indicating the catch and release mortality of recreational fishery. This \cr
is the change that fish dies after being landed and released in a recreational fishery. The \cr
default value is 0.14. \if{html}{\out{<Source>}}}

\item{bootstrap}{Boolean indicating if parametric bootstrapping should be conducted. If \code{bootstrap} \cr
is set to false, then \code{iter} is ignored.}

\item{iter}{Numeric or integer specifying the number of iterations for parametric bootstrapping. The \cr
default value is 1000. Note that larger values require more computing power and memory.}

\item{detail}{Boolean indicating if the reconstruction should calculate breakdowns and summary \cr
information.}

\item{min_harvest_rate}{Double specifying the lower bound for the harvest rate considered in cohort \cr
reconstruction. Harvest rate is the ratio between the number of harvestable individuals and the \cr
size of the cohort.}

\item{level}{Double between 0 and 1 inclusive, specifying the credible level for credible intervals. \cr
This option only matters if \code{bootstrap} is set to TRUE.}

\item{hpd}{Boolean indicating if the highest posterior density credible interval should be used. If \cr
FALSE, a symmetric credible interval is used.}

\item{verbose}{Boolean indicating if the function will print progress updates. If \code{verbose} is set \cr
to TRUE, the function will provide status updates. This is particularly useful for conducting \cr
bootstrapped cohort reconstruction for large iterations that are high in computing cost, as \cr
having progress updates can instill a sense of confidence into an otherwise discouraged user.}

\item{sex}{String specifying which sex or sexes to consider. Must choose from "male", "female", or \cr
"both". Here, "male" is shorthand for sperm-producing individuals. "Female" is shorthand for \cr
egg-producing individuals.}
}
\value{
A three-dimensional list of results. The first dimension encodes brood year information (by), \cr
the second age, and the third month. Each age has an age-specific summary information. Each month has \cr
\code{data} and month-specific \code{summary} information.Depending on \code{detail} and \code{bootstrap}, the corresponding \cr
output will vary. Here is the structure for the return.

\if{html}{\out{<div class="sourceCode">}}\preformatted{  "by" "age" "month"

├── [["2002"]]
│   ├──[["2"]]
│   │   ├── [["1"]]
│   │   │   ├── `data`
│   │   │   └── `summary`
│   │   ├── ...
│   │   │
│   │   ├── [["age_summary"]]
│   │   │   ├── `data`
│   │       └── `summary`
│   ├── [["3"]]
│   │   ├── [["1"]]
│   │   │   ├── `data`
│   │   │   └── `summary`
│   │   ├── ...
│   │   │
│   │   ├── [["age_summary"]]
│   │   │   ├── `data`
│   │       └── `summary`
│   ├── ...
│   │ 
│   ├──[["by_summary"]]
│   │  ├── `data`
│   │  └── `summary`
├── ...

}\if{html}{\out{</div>}}
\subsection{Summary}{
\subsection{Brood-year-specific summary (under \code{by_summary})}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Parameter | Median | SD  | CrI_low | CrI_high |
|-----------|--------|-----|---------|----------|
| srr       | 0.05   | 0.02| 0.04    | 0.04     |
}\if{html}{\out{</div>}}
}

\subsection{Age-specific summary (under \code{age_summary})}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Parameter | Median | SD  | CrI_low | CrI_high |
|-----------|--------|-----|---------|----------|
| elsr      | 0.02   | 0   | 0.01    | 0.01     |
}\if{html}{\out{</div>}}
}

\subsection{Month-specific summary (under each month)}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Parameter         | Median | SD   | CrI_low | CrI_high |
|-------------------|--------|------|---------|----------|
| ocean_abundance   | 43.91  | 6.50 | 35.09   | 35.09    |
| impact            | 0.00   | 0.00 | 0.00    | 0.00     |
| maturation        | 0.00   | 0.00 | 0.00    | 0.00     |
| natural_mort      | 0.81   | 0.12 | 0.65    | 0.65     |
}\if{html}{\out{</div>}}
}

}

\subsection{Data}{
\subsection{Brood-year-specific data}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Parameter | 1    | 2    | 3    | 4    | 5    | 6    | 7    | 8    | 9    | 10   |
|-----------|------|------|------|------|------|------|------|------|------|------|
| srr       | 0.07 | 0.07 | 0.05 | 0.05 | 0.05 | 0.06 | 0.06 | 0.04 | 0.08 | 0.04 |

}\if{html}{\out{</div>}}
}

\subsection{Age-specific data}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Measurement | 1    | 2    | 3    | 4    | 5    | 6    | 7    | 8    | 9    | 10   |
|-------------|------|------|------|------|------|------|------|------|------|------|
| elsr        | 0.02 | 0.02 | 0.02 | 0.01 | 0.01 | 0.02 | 0.02 | 0.02 | 0.01 | 0.02 |
}\if{html}{\out{</div>}}
}

\subsection{Month-specific data}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Measurement      | 1    | 2    | 3    | 4    | 5    | 6    | 7    | 8    | 9    | 10   |
|------------------|------|------|------|------|------|------|------|------|------|------|
| ocean_abundance  | 46.69| 37.41| 43.34| 40.74| 35.09| 44.49| 37.43| 56.25| 45.80| 50.56|
| impact           | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 |
| maturation       | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 | 0.00 |
| natural_mort     | 0.86 | 0.69 | 0.80 | 0.75 | 0.65 | 0.82 | 0.69 | 1.04 | 0.84 | 0.93 |
}\if{html}{\out{</div>}}
}

}
}
\description{
\code{cohort_reconstruct} uses released and recovered coded wire tags to calculate the ocean \cr
abundance of the population for each brood year at each age and in each month. The function \cr
returns a list of outputs indexed by brood year, age, and month. Each element of the list \cr
contains two elements, \code{data} and \code{summary}. In general, the \code{data} segment contains the \cr
raw data of cohort reconstruction, while the \code{summary} segment includes statistics from the \cr
raw data. For more about return value, see the value section. \cr

Depending on if the \code{bootstrap} and \code{detail} flags are set, the \code{data} and \code{summary} segments \cr
may behave in different ways. For more information, see details.
}
\details{
\subsection{Aging convention}{

We use fishing age, or year of life, as the age for captured tags. This aging convention \cr
starts from age-1. Namely, as soon as the fish is born, it is in the first year of life, \cr
thus age-1. Accordingly, the age in the \code{size_at_age} data frame should use fishing age \cr
as well. \cr
}

\subsection{Detail, bootstrap, and results}{

Since \code{detail} and \code{bootstrap} are Boolean valued variables, there are four combinations.
\describe{
\item{\verb{detail = True, bootstrap = True}}{Provides results with detailed summary statistics using bootstrapped data.}
\item{\verb{detail = False, bootstrap = True}}{Provides results with cohort data only using bootstrapped data.}
\item{\verb{detail = True, bootstrap = False}}{Provides results with detailed summary statistics using point estimates.}
\item{\verb{detail = False, bootstrap = True}}{Provides results with cohort data only using point estimates.}
}
}

\subsection{Danger of incomplete cohort reconstruction}{

If the provided recovery data don't span the entire life time of a stock, cohort reconstruction \cr
may fail to yield accurate information. Be sure to provide a complete cohort recovery data set.
}
}
\examples{
cohort_reconstruct(release, recovery, birth_month = 4,
  bootstrap = F, detail = F, last_month = 12L, iter = 10,
  level = 0.95, , sex = "both")
}
