% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_at_age}
\alias{length_at_age}
\title{Length-at-age information}
\format{
\subsection{\code{length_at_age}}{

\describe{
\item{"age"}{Integer. The fishing age of the cohort.}
\item{"month"}{Integer. The month in numeric form. For instance, \code{3L} is March.}
\item{"mean"}{Double. The mean total body length in inches.}
\item{"sd"}{Double. The standard deviation of total body length in inches.}
}
}
}
\usage{
length_at_age
}
\description{
length-at-age estimated in \insertCite{age_struct}{cwtCensus}. Estimated \cr
using the Basic Model described in \insertCite{bm}{cwtCensus} from hatchery fish from \cr
2000-2020 originating from Coleman National Fish Hatchery, Feather River Hatchery, and \cr
Nimbus Fish Hatchery.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
