% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_at_age}
\alias{length_at_age}
\title{Length at age, or key length age, information}
\format{
\subsection{'release'}{

\describe{
\item{age}{The age of fish in which the CWT is recovered}
\item{month}{The month in which the CWT is recovered}
\item{mean}{The mean legnth of fish in inches for the specified age and month}
\item{sd}{The standard deviation of the fish length in inches for the specified age and month}
}
}
}
\usage{
length_at_age
}
\description{
Length at age, or key length age, information
}
\keyword{datasets}
