% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recovery}
\alias{recovery}
\title{Recovery information}
\format{
\subsection{\code{recovery}}{

\describe{
\item{"run_year"}{Integer. The year of each CWT recovery.}
\item{"month"}{Integer. The month of each CWT recovery.}
\item{"length"}{Double. The length in inches of each recovered fish.}
\item{"tag_code"}{Any type. The identifier for each recovered tag.}
\item{"fishery"}{Double. The identifier for the fishery from which the tag is recovered. \cr
The fisheries include river harvest, spawning ground escapement, hatchery \cr
escapement, ocean recreational fishing, ocean commercial fishing.}
\item{"location"}{Character. The identifier for the location at which the tag is recovered.}
\item{"size_limit"}{Double. The minimum harvest limit (inclusive) in inches for the recovered \cr
tag's region and date.}
\item{"est_num"}{Double. The estimated number of fish each tag represents.}
\item{"sex"}{Character. The biological sex of the recovered fish.}
}
}
}
\usage{
recovery
}
\description{
Recovery information
}
\keyword{datasets}
