% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{release}
\alias{release}
\title{Release information}
\format{
\subsection{\code{release}}{

\describe{
\item{"brood_year"}{Integer. The birth year of each CWT batch.}
\item{"total_release"}{Double. The total number of releases for each batch of CWT.}
\item{"prod_exp"}{Double. The ratio between the number of tagged fish and the total number \cr
of fish in each batch, i.e., number of tagged fish / batch size.}
\item{"tag_code"}{Any type. The identifier for each batch of CWT release.}
}
}
}
\source{
Regional Mark Processing Center (RMPC), Pacific States Marine Fisheries Commission (2024) \cr
Regional Mark Information System Database. www.rmpc.org (accessed 31 Jan 2024)
}
\usage{
release
}
\description{
Nimbus Fish Hatchery release information 2007 - 2009.
}
\keyword{datasets}
