% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{release_mort}
\alias{release_mort}
\title{Release mortality information}
\format{
\subsection{\code{release_mort}}{

\describe{
\item{"run_year"}{Integer. The year of each CWT recovery.}
\item{"fishery"}{Any type. The identifier for each type of fishery.}
\item{"location"}{Any type. The identifier for location.}
\item{"month"}{Integer. The month of each CWT recovery.}
\item{"rate"}{Double. The release mortality rate for each timestep.}
}
}
}
\usage{
release_mort
}
\description{
hook-and-release mortality described in \insertCite{age_struct}{cwtCensus}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
